Add OS other than Windows support
Add OS other than Windows support

+++ src/dxvk/dxvk_bind_mask.h
@@ -144,7 +144,7 @@ namespace dxvk {
       if (!mask)
         return -1;
       
-      return BitCount * intId + bit::tzcnt(mask);
+      return BitCount * intId + (int32_t)bit::tzcnt((uint32_t)mask);
     }
 
     bool operator == (const DxvkBindingSet& other) const {
